+++ src/pages/api/contact.tsx
@@ -16,7 +16,15 @@ export default async (req: NextApiRequest, res: NextApiResponse) => {
     return;
   }
 
-  const { firstName, lastName, email, organization, role, message } = req.body;
+  const {
+    firstName,
+    lastName,
+    email,
+    organization,
+    role,
+    message,
+    url,
+  } = req.body;
 
   await fetch(SLACK_WEBHOOK_CONTACT_US, {
     method: "POST",
@@ -28,7 +36,8 @@ export default async (req: NextApiRequest, res: NextApiResponse) => {
         `*Email*: <mailto:${email}|${email}>\n` +
         `*Organization*: ${organization}\n` +
         `*Role*: ${role}\n` +
-        `*Message*: ${message}`,
+        `*Message*: ${message}` +
+        `*URL*: ${url}`,
     }),
   });
